PR-H round-2 Copilot review fixes
Two valid issues from Copilot.

1. tests-cpp/small/test_capi_i64.cpp: gated test was scoped too small.
   HUGE_N = 2200 * 1024 * 1024 = ~2.3 GiB, which is < UINT32_MAX (~4.29
   GiB). As written, the test would not have caught a regression where
   the _i64 entries were accidentally wired through uint32 narrowing --
   the resize / writes / reads would all land in the low-32-bit-
   addressable region.

   Bumped HUGE_N to 5 GiB (uint64_t(5) * 1024 * 1024 * 1024), gated by
   static_assert that the value exceeds UINT32_MAX. Sentinel writes /
   reads now happen at IDX_A = UINT32_MAX + 17 (just past the 32-bit
   boundary), IDX_B = HUGE_N - 2, IDX_C = HUGE_N - 1 -- all comfortably
   past UINT32_MAX.

2. src/misc/daScriptC.cpp: das_array_at_i64 pointer arithmetic used
   uint64_t directly:

       return ((Array *)arr)->data + uint64_t(index) * uint64_t(stride);

   The legacy das_array_at uses size_t. On 64-bit platforms size_t ==
   uint64_t so the bits match; on 32-bit platforms size_t is uint32_t
   and char* + uint64_t would trigger an implicit narrowing
   (-Wconversion) -- and a huge index can't be addressed on 32-bit
   anyway, same graceful degradation the legacy entry already provides.

   Switched to size_t to match das_array_at:

       return ((Array *)arr)->data + size_t(index) * size_t(stride);

Verification:
- tests-cpp-small: 39/39.
- DASLANG_HUGE_HEAP_TESTS=1: huge-resize SUBCASE still passes with the
  bumped HUGE_N + indices > UINT32_MAX.

+++ src/misc/daScriptC.cpp
@@ -1193,7 +1193,11 @@ void * das_array_at ( das_array * arr, uint32_t index, uint32_t stride ) {
 }
 
 void * das_array_at_i64 ( das_array * arr, uint64_t index, uint32_t stride ) {
-    return ((Array *)arr)->data + uint64_t(index) * uint64_t(stride);
+    // size_t matches das_array_at's pattern: 64-bit on 64-bit platforms (where
+    // arr.size > UINT32_MAX is reachable), 32-bit on 32-bit platforms (where
+    // a huge index can't be addressed anyway -- same graceful degradation as
+    // the legacy entry).
+    return ((Array *)arr)->data + size_t(index) * size_t(stride);
 }
 
 void das_array_lock ( das_context * context, das_array * arr ) {
